Add support for interpreting SVG presentation attributes:
Support is added for the SVG 1.1 spec's set of presentation attributes.
Many of these do not exist as WL options and refer to XML elements that
are specific to the SVG namespace (not HTML elements). For that reason
we parse the attributes as key-value pairs using keys prefixed with
"SVG" to avoid ambiguity with any future WL option names.

To put the SVG attribute data in line with the CSS property data we add
new property keys "Animatable", "AppliesTo", and "Values" to both data
structures. "AppliesTo" makes more sense for the SVG namespace but it
also works for general CSS properties. However for CSS it is harder to
establish restrictions since CSS can apply to all XML and established
namespaces.