ValuesAndUnits3 - Bugfixes for calc():
Parenthetical nesting was causing issues, so I converted calcReduce to
be recursive. Whenever a term or product is detected as either calc() or
(), then we first reduce the containing token as far as possible before
continuing the calculation.

To this end we had to put in the case of only a single term e.g. (3) or
(  4px ) during any addition step.