Images3 - work in progress
I decided to expose just two functions, CSSLinearGradient and
CSSRadialGradient, instead of the many functions that convert
individual parts. If I am already required to use a special
function to make CSS-compliant gradients that requires the CSSTools
package to be loaded in the kernel, then there's no reason to bend
over backwards to make the exposed function look like
LinearGradientImage and RadialGradientImage. They are still used
to generate the images, but are now simply implementation details.

Other changes:
* use Automatic instead of None in the exposed functions
* improved nonlinear gradient hint calculation that uses Plot to sample points
* finished radial gradient implementation (much simplified code refactor)
* use WL canonicalized strings e.g. "ClosestCorner" instead of CSS "closest-corner"
* corrections to linear gradient implementation
* resuse code for repeating gradients, but with option "Repeating"->True
* repeating gradients required recalculation for linear corner-to-corner gradients