Images3 - completion of coding
Only two functions remain exposed: CSSLinearGradientImage and
CSSRadialGradientImage. The slight rename is to follow the existing
functions within WD (...GradientImage). I settled on a syntax for
the functions that follows more of the CSS style than the WD style. The
shape and size (I call 'extent') arguments for radial gradients were
moved to options. The repeating and non-repeating versions of the CSS
functions remain merged into just two WD functions with a "Repeating"
option.

Following WD conventions for ...GradientImage functions, 0 to 3 argument
versions exist for each CSS...GradientImage function. This is also in
line with CSS which only requires the color stop list as a minimum
syntax. A grayscale gradient is used if no color stop list is provided.

One of the biggest changes is in fixing the implementation of repeating
gradients to match the CSS spec. In effect, a repeating gradient is the
same as a non-repeating one, but with the color stop list repeated.
Before, I had tried to rely on the WD option Padding->"Periodic", but
this was insufficient due to the interaction between color stop
positions and the calculated length of the gradient ray. Instead, I
manually duplicate or pad the color stop list as needed. It can be more
computationally intensive for many closely placed color stops, but it is
faithful to the CSS spec.

Summary of changes:
* refactor of parsing code using a isRepeating switch
* moved code blocks around and added section, subsection etc.
* fixed implementation of repeating gradients to match CSS spec
* added internal documentation about repeating gradients
* other small adjustments