Reorganize project layout:
The Eclipse project is reorganized to be cleaner and modernized. The
subdirectory CSSTools is now the fully realized paclet structure.

You can no longer "run" notebooks from Eclipse, but this is not a loss.
Instead, to test any functionality while actively working on the
project, open a Mathematica or WD session (version 12.1 or later) and
load the paclet directly via

  PacletDirectoryLoad[<path to eclipse project>]

then evaluate Get["CSSTools`"] to load the definitions. If any changes
occur to the source code, such as by editing it within the Eclipse IDE,
then run Get again to update the definitions.

The documentation is still built with Eclipse. The doc templates are
found within WLDocumentationTemplates and the built docs are directly
built into the paclet layout. The build script is updated with the new
paths so the workflow remains the same: edit the templates and run the
build script as usual.

Summarizing the changes:

Paclet Structure
* moved paclet info file /PacletInfo.m --> /CSSTools/PacletInfo.m
* moved WL code /CSSTools/*.wl --> /CSSTools/Kernel/*.wl
* set kernel root in PacletInfo.m to "Kernel"
* deleted init.m file
* edited docbuild.xml to build doc pages directly to /CSSTools/Documentation

Documentation Build Process
* moved WL documentation templates /CSSTools/Documentation --> /WLDocumentationTemplates
* edited docbuild.xml to build formatted docs directly into paclet layout
* docbuild.xml also deletes built docs before rebuilding them

Unit Testing
* moved unit tests /CSSTools/Testing --> /Testing
* moved scratch notebook /CSSTools/CSSTools.nb --> /Testing/CSSTools.nb

Cleanup
* deleted /CSSTools/Old
* moved markdown files /*.md --> /ReadMeFiles/*.md (except README.md)